requirements.txt
--index-url https://pypi.org/simple/
--extra-index-url https://test.pypi.org/simple/

tigrmail==0.3.0
beautifulsoup4==4.12.3
playwright==1.48.0
pytest==8.3.3
pytest-playwright==0.5.1
python-dotenv==1.0.1
